requirements.txt
pandas
python-dotenv
numpy
requests
matplotlib
seaborn
plotly
sqlalchemy
lxml
psycopg2-binary==2.9.3
